+++ tests/test_visitors/test_ast/test_classes/test_methods/test_async_magic_methods.py
@@ -1,0 +1,104 @@
+# -*- coding: utf-8 -*-
+
+import pytest
+
+from wemake_python_styleguide.constants import ASYNC_MAGIC_METHODS_WHITELIST
+from wemake_python_styleguide.violations.oop import AsyncMagicMethodViolation
+from wemake_python_styleguide.visitors.ast.classes import WrongMethodVisitor
+
+sync_method = """
+class Example(object):
+    def {0}(self): ...
+"""
+
+async_method = """
+class Example(object):
+    async def {0}(self): ...
+"""
+
+
+@pytest.mark.parametrize('method', [
+    '__init__',
+    '__eq__',
+    '__lt__',
+    '__next__',
+    '__enter__',
+    '__exit__',
+])
+def test_wrong_async_magic_used(
+    assert_errors,
+    assert_error_text,
+    parse_ast_tree,
+    method,
+    default_options,
+):
+    """Testing that some async magic methods are restricted."""
+    tree = parse_ast_tree(async_method.format(method))
+
+    visitor = WrongMethodVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [AsyncMagicMethodViolation])
+    assert_error_text(visitor, method)
+
+
+@pytest.mark.parametrize('method', ASYNC_MAGIC_METHODS_WHITELIST)
+def test_correct_async_magic_used(
+    assert_errors,
+    parse_ast_tree,
+    method,
+    default_options,
+):
+    """Testing that some async magic methods are working fine."""
+    tree = parse_ast_tree(async_method.format(method))
+
+    visitor = WrongMethodVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [])
+
+
+@pytest.mark.parametrize(
+    'method',
+    ASYNC_MAGIC_METHODS_WHITELIST | {
+        '__init__',
+        '__eq__',
+        '__lt__',
+        '__next__',
+        '__enter__',
+        '__exit__',
+    },
+)
+def test_sync_magic_used(
+    assert_errors,
+    parse_ast_tree,
+    method,
+    default_options,
+):
+    """Testing that any sync magic methods are working fine."""
+    tree = parse_ast_tree(sync_method.format(method))
+
+    visitor = WrongMethodVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [])
+
+
+@pytest.mark.parametrize('method', [
+    'next',
+    'regular',
+])
+def test_regular_method_used(
+    assert_errors,
+    parse_ast_tree,
+    method,
+    mode,
+    default_options,
+):
+    """Testing that other methods are working fine."""
+    tree = parse_ast_tree(mode(sync_method.format(method)))
+
+    visitor = WrongMethodVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [])